Added repository

+++ tdd-java8-jbehave/src/main/java/com/weebly/gaborcsikos/tdd_java8_jbehave/apple/AppleController.java
@@ -26,7 +26,7 @@ public class AppleController {
 
 	}
 
-	public void deleteById(int appleId) {
+	public void deleteById(long appleId) {
 		model.removeById(appleId);
 
 	}
