+++ tdd-java8-jbehave/src/main/java/com/weebly/gaborcsikos/tdd_java8_jbehave/apple/AppleModel.java
@@ -18,7 +18,7 @@ public class AppleModel {
 		return apples;
 	}
 
-	public void removeById(int appleId) {
+	public void removeById(long appleId) {
 		apples.removeIf(p -> p.getID() == appleId);
 
 	}
