+++ tdd-java8-jbehave/src/test/java/com/weebly/gaborcsikos/tdd_java8_jbehave/unit/AppleControllerTest.java
@@ -26,7 +26,7 @@ import com.weebly.gaborcsikos.tdd_java8_jbehave.apple.AppleController;
 @RunWith(MockitoJUnitRunner.class)
 public class AppleControllerTest {
 
-	private static final int APPLE_ID = 1;
+	private static final long APPLE_ID = 1;
 
 	@InjectMocks
 	private AppleController appleController;
